employees_service/requirements.txt
fastapi>=0.87.0
uvicorn>=0.19.0
# python>=3.11
sqlalchemy>=1.4.44
alembic==1.8.1
psycopg2-binary>=2.9.5
retrying>=1.3.3
secure>=0.3.0
passlib>=^1.7.4
python-jose>=3.3.0
python-multipart>=0.0.5
bcrypt>=4.0.1
pytest>=7.2.0
requests>=2.28.1
coverage>=6.5.0
httpx>=0.23.1